Split out Types of Values in Assignments
